gcp/deltasharing/consumer/examples/python/requirements.txt
requests
uuid
jwcrypto
google-cloud-storage
google-auth